add profile feature

+++ Spring-boot-tutorial-Nhut-context-root/src/main/java/com/caveofprogramming/controllers/ProfileController.java
@@ -16,12 +16,14 @@ import org.owasp.html.PolicyFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.InputStreamResource;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -37,6 +39,7 @@ import com.caveofprogramming.model.SiteUser;
 import com.caveofprogramming.service.FileService;
 import com.caveofprogramming.service.ProfileService;
 import com.caveofprogramming.service.UserService;
+import com.caveofprogramming.status.PhotoUploadStatus;
 
 /**
  * @author java_dev
@@ -59,15 +62,31 @@ public class ProfileController {
 	@Value("${photo.upload.directory}")
 	private String photoUploadDirectory;
 	
+	@Value("${photo.upload.ok}")
+	private String photoUploadOk;
+	
+	@Value("${photo.upload.invalid}")
+	private String photoUploadInvalid;
+	
+	@Value("${photo.upload.error}")
+	private String photoUploadError;
+	
+	@Value("${photo.upload.toosmall}")
+	private String photoUploadTooSmall;
+	
 	private SiteUser getUser() {
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		String username = auth.getName();
 		return userService.get(username);
 	}	
 	
-	@RequestMapping(value="/profile")
-	public ModelAndView showProfile(ModelAndView modelAndView) {		
-		SiteUser user = getUser();
+	private ModelAndView showProfile(SiteUser user) {
+		ModelAndView modelAndView = new ModelAndView();
+		if (user == null) {
+			modelAndView.setViewName("redirect:/");
+			return modelAndView;
+		}
+		
 		Profile profile = profileService.getUserProfile(user);
 		if (profile == null) {
 			profile = new Profile();
@@ -76,9 +95,28 @@ public class ProfileController {
 		}
 		Profile webProfile = new Profile();
 		webProfile.safeCopyFrom(profile);
-		
+
+		modelAndView.getModel().put("userId", user.getId());
 		modelAndView.getModel().put("profile", webProfile);		
 		modelAndView.setViewName("app.profile");
+		return modelAndView;		
+	}
+	
+	@RequestMapping(value="/profile")
+	public ModelAndView showProfile() {		
+		SiteUser user = getUser();
+		
+		ModelAndView modelAndView = showProfile(user);
+		
+		return modelAndView;
+	}
+	
+	@RequestMapping(value="/profile/{id}")
+	public ModelAndView showProfile(@PathVariable("id") Long id) {	
+		SiteUser user = userService.get(id);
+		
+		ModelAndView modelAndView = showProfile(user);
+		
 		return modelAndView;
 	}
 	
@@ -114,9 +152,10 @@ public class ProfileController {
 	}
 	
 	@RequestMapping(value="/upload-profile-photo", method=RequestMethod.POST)
-	public ModelAndView handlePhotoUploads(ModelAndView modelAndView, 
-			@RequestParam("file") MultipartFile file) throws IOException, InvalidFileException, ImageTooSmallException {
-		modelAndView.setViewName("redirect:/profile");
+	@ResponseBody // Return data in JSON format
+	public ResponseEntity<PhotoUploadStatus> handlePhotoUploads(/*ModelAndView modelAndView,*/ 
+			@RequestParam("file") MultipartFile file) {
+//		modelAndView.setViewName("redirect:/profile");
 //		Path outputFilePath = Paths.get(photoUploadDirectory, file.getOriginalFilename());
 //		Files.deleteIfExists(outputFilePath);
 //		Files.copy(file.getInputStream(), outputFilePath);
@@ -124,20 +163,53 @@ public class ProfileController {
 		SiteUser user = getUser();
 		Profile profile = profileService.getUserProfile(user);
 		Path oldPhotoPath = profile.getPhoto(photoUploadDirectory);
-		FileInfo photoInfo = fileService.saveImageFile(file, photoUploadDirectory, "photos", "p" + user.getId(), 100, 100);		
-		profile.setPhotoDetails(photoInfo);
-		profileService.save(profile);
 		
-		if (oldPhotoPath != null)
-			Files.delete(oldPhotoPath);
+		PhotoUploadStatus status = new PhotoUploadStatus(photoUploadOk);
+		
+		try {
+			FileInfo photoInfo = fileService.saveImageFile(file, photoUploadDirectory, "photos", "p" + user.getId(), 100, 100);		
+			profile.setPhotoDetails(photoInfo);
+			profileService.save(profile);
+			
+			if (oldPhotoPath != null) Files.delete(oldPhotoPath);
+		} catch (InvalidFileException e) {
+			status.setMessage(photoUploadInvalid);
+			e.printStackTrace();
+		} catch (IOException e) {
+			status.setMessage(photoUploadError);
+			e.printStackTrace();
+		} catch (ImageTooSmallException e) {
+			status.setMessage(photoUploadTooSmall);
+			e.printStackTrace();
+		}
 				
-		return modelAndView;
+//		return modelAndView;
+//		return status;
+		return new ResponseEntity(status, HttpStatus.OK);
 	}
 	
 	@RequestMapping(value="/profilephoto", method=RequestMethod.GET)
 	@ResponseBody
 	public ResponseEntity<InputStreamResource> servePhoto() throws IOException {
 		SiteUser user = getUser();
+		Profile profile = profileService.getUserProfile(user);
+		
+		Path photoPath = Paths.get(photoUploadDirectory, "default", "avatar.png");
+		if(profile != null && profile.getPhoto(photoUploadDirectory) != null) {
+			photoPath = profile.getPhoto(photoUploadDirectory);
+		}
+		
+		return ResponseEntity
+			.ok()
+			.contentLength(Files.size(photoPath))
+			.contentType(MediaType.parseMediaType(URLConnection.guessContentTypeFromName(photoPath.toString())))
+			.body(new InputStreamResource(Files.newInputStream(photoPath, StandardOpenOption.READ)));
+	}
+	
+	@RequestMapping(value="/profilephoto/{id}", method=RequestMethod.GET)
+	@ResponseBody
+	public ResponseEntity<InputStreamResource> servePhoto(@PathVariable Long id) throws IOException {
+		SiteUser user = userService.get(id);
 		Profile profile = profileService.getUserProfile(user);
 		
 		Path photoPath = Paths.get(photoUploadDirectory, "default", "avatar.png");
